🎨 Split Notion client and fetch functions into separate files
